backend/requirements.txt
asgiref==3.10.0
Django==5.2.7
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.0
sqlparse==0.5.3
tzdata==2025.2
pandas
openpyxl
google-generativeai
python-dateutil==2.8.2

# Cache e Performance
redis==5.0.1
django-redis==5.4.0

# Processamento Assíncrono
celery==5.3.6
celery[redis]==5.3.6

# Rate Limiting
django-ratelimit==4.1.0

# File Validation
# python-magic removed - using magic bytes validation instead (Windows compatible)

# API Documentation
drf-spectacular==0.27.0
